Bring up to date with JUCE 6 and C++17
This project won't compile anymore. It took some investigation of the core C++ spec to figure out the right thing to do with some of the JUCE parameter calls that are not backward-compatible, but the code has been tested and debugged to verify the changes made were the correct ones.